Move CreateCompactionJobs to compaction-core module
