+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/creation/AssignJobIdQueueSender.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.creation.commit;
+package sleeper.compaction.core.job.creation;
 
 import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequest;
