+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/job/creation/strategy/impl/CompactionStrategyTestBase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.strategy.impl;
+package sleeper.compaction.core.job.creation.strategy.impl;
 
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobFactory;
