+++ java/compaction/compaction-job-creation/src/test/java/sleeper/compaction/job/creation/CreateCompactionJobsIT.java
@@ -38,7 +38,9 @@ import sleeper.compaction.core.job.CompactionJobSerDe;
 import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequest;
 import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequestSerDe;
-import sleeper.compaction.job.creation.CreateCompactionJobs.Mode;
+import sleeper.compaction.core.job.creation.CreateCompactionJobs;
+import sleeper.compaction.core.job.creation.CreateCompactionJobs.Mode;
+import sleeper.compaction.core.job.creation.CreateJobsTestUtils;
 import sleeper.configuration.properties.S3InstanceProperties;
 import sleeper.configuration.table.index.DynamoDBTableIndexCreator;
 import sleeper.core.CommonTestConstants;
@@ -63,9 +65,9 @@ import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequestTestHelper.requestToAssignFilesToJobs;
-import static sleeper.compaction.job.creation.CreateJobsTestUtils.assertAllReferencesHaveJobId;
-import static sleeper.compaction.job.creation.CreateJobsTestUtils.createInstanceProperties;
-import static sleeper.compaction.job.creation.CreateJobsTestUtils.createTableProperties;
+import static sleeper.compaction.core.job.creation.CreateJobsTestUtils.assertAllReferencesHaveJobId;
+import static sleeper.compaction.core.job.creation.CreateJobsTestUtils.createInstanceProperties;
+import static sleeper.compaction.core.job.creation.CreateJobsTestUtils.createTableProperties;
 import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.COMPACTION_JOB_QUEUE_URL;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.CONFIG_BUCKET;
