+++ java/core/src/main/java/sleeper/core/properties/instance/CompactionProperty.java
@@ -267,9 +267,9 @@ public interface CompactionProperty {
             .propertyGroup(InstancePropertyGroup.COMPACTION).build();
     UserDefinedInstanceProperty DEFAULT_COMPACTION_STRATEGY_CLASS = Index.propertyBuilder("sleeper.default.compaction.strategy.class")
             .description("The name of the class that defines how compaction jobs should be created. " +
-                    "This should implement sleeper.compaction.core.strategy.CompactionStrategy. The value of this property is the " +
+                    "This should implement sleeper.compaction.core.job.creation.strategy.CompactionStrategy. The value of this property is the " +
                     "default value which can be overridden on a per-table basis.")
-            .defaultValue("sleeper.compaction.core.strategy.impl.SizeRatioCompactionStrategy")
+            .defaultValue("sleeper.compaction.core.job.creation.strategy.impl.SizeRatioCompactionStrategy")
             .propertyGroup(InstancePropertyGroup.COMPACTION).build();
     UserDefinedInstanceProperty DEFAULT_COMPACTION_FILES_BATCH_SIZE = Index.propertyBuilder("sleeper.default.compaction.files.batch.size")
             .description("The maximum number of files to read in a compaction job. Note that the state store must " +
